Exception handling

+++ dscommerce/src/main/java/com/oarthurfc/dscommerce/services/ProductService.java
@@ -1,14 +1,20 @@
 package com.oarthurfc.dscommerce.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.dao.DataIntegrityViolationException;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Propagation;
 import org.springframework.transaction.annotation.Transactional;
 
 import com.oarthurfc.dscommerce.dto.ProductDTO;
 import com.oarthurfc.dscommerce.entities.Product;
 import com.oarthurfc.dscommerce.repositories.ProductRepository;
+import com.oarthurfc.dscommerce.services.exceptions.DatabaseException;
+import com.oarthurfc.dscommerce.services.exceptions.ResourceNotFoundException;
+
+import jakarta.persistence.EntityNotFoundException;
 
 @Service
 public class ProductService {
@@ -25,7 +31,8 @@ public class ProductService {
 
     @Transactional(readOnly = true)
     public ProductDTO findById(Long id){
-        Product product = repository.findById(id).get();
+    Product product = repository.findById(id).orElseThrow(
+        () -> new ResourceNotFoundException("Elemento não encontrado"));
         return new ProductDTO(product);
     }
 
@@ -37,15 +44,26 @@ public class ProductService {
 
     @Transactional
     public ProductDTO update(Long id, ProductDTO dto){
-        Product product = repository.getReferenceById(id);
-        copyDtoToEntity(dto, product);
-        product = repository.save(product);
-        return new ProductDTO(product);
+        try {
+            Product product = repository.getReferenceById(id);
+            copyDtoToEntity(dto, product);
+            product = repository.save(product);
+            return new ProductDTO(product);
+        } catch (EntityNotFoundException e) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
     }
 
-    @Transactional
+    @Transactional(propagation = Propagation.SUPPORTS)
     public void delete(Long id){
-        repository.deleteById(id);
+        if (!repository.existsById(id)) {
+            throw new ResourceNotFoundException("Recurso não encontrado");
+        }
+        try {
+            repository.deleteById(id);
+        } catch (DataIntegrityViolationException e){
+            throw new DatabaseException("Falha de integridade referencial");
+        }
     }
         
     private void copyDtoToEntity(ProductDTO dto, Product product) {
